Test refactoring / abstraction
No functional change.

This is a preparation PR for the upcoming addition of global-states, which involves quite a bit of test refactorings.
