+++ servers/jax-rs-tests/src/test/java/org/projectnessie/jaxrs/TestJerseyResteasy.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2020 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.jaxrs;
+
+import io.restassured.RestAssured;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.http.ContentType;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.extension.RegisterExtension;
+
+public class TestJerseyResteasy extends AbstractResteasyTest {
+
+  @RegisterExtension static NessieJaxRsExtension server = new NessieJaxRsExtension();
+
+  @BeforeAll
+  static void setup() {
+    RestAssured.baseURI = server.getURI().toString();
+    RestAssured.port = server.getURI().getPort();
+    RestAssured.requestSpecification =
+        new RequestSpecBuilder()
+            .setContentType(ContentType.JSON)
+            .setAccept(ContentType.JSON)
+            .build();
+    basePath = "/";
+  }
+}
