+++ servers/quarkus-server/src/test/java/org/projectnessie/server/ITNativeBasicOperations.java
@@ -18,4 +18,4 @@ package org.projectnessie.server;
 import io.quarkus.test.junit.NativeImageTest;
 
 @NativeImageTest
-public class ITNativeBasicOperations extends TestBasicOperations {}
+public class ITNativeBasicOperations extends AbstractTestBasicOperations {}
